User Validation in both MainForm and RTControlForm.
Didn't git-add these files before, so here they are! Once again,
all user input in both forms has been validated and tested.
Implemented the Validator class, which contains methods to
do these checks for us. Fixed Speed box in RTControlForm, as well
as the issue of the caseswitch statement not matching the index ID
of the dropdown lists that were created. Issue #183