Config File implemented & tested
Added the JSON Config File used to allow the control room software to determine which telescope to run. You are now able to pick an existing RT ID from the database and run it (if it exists).
Made some changes to the MainForm.cs logic to account for error checking and the new functionality.