Added error handling for JSON file being corrupted
User will now be notified if there was an error parsing the JSON file (corrupted, invalid data, etc.) and then ask the user if they would like to recreate it.
Message boxes will display to remind the user that they need to go into the file and change the ID to the selected telescope they would like to run, as well as to revert the newTelescope flag back to false.
Added a DEFAULT_JSON_CONTENTS string that can be modified if we decide to change what the default should be in the future.
Added a helper method, CreateAndWriteToNewJSONFile, that takes care of writing to the file, and recreating it if it does not exist.