Error handling for FileNotFoundException
There was a case in which a user who did not have the JSON file created would run into an exception upon the program trying to read from the RadioTelescopeConfig.json file.
That issue has been fixed, and the program will now properly recreare the file if it does not exist.