Updated coordinate calculation to NOT modify curr orientation
Previously, the coordinate calculation was modifying the current
orientation. This isn't problematic unless our orientation is stored in
a variable, but regardless, it's a good precaution to NOT do this.
C# objects are passed by value, but their inner variables are passed by
reference, so they can still be modified.

Issue #250