Simulation now keeps track of the telescope's position
This is a TEMPORARY fix that was done because we can't find the
simulation's source code. What we should be doing is having the
simulation send the position, and the control room reads it (similar to
what the encoders are doing in the MCU), but we don't have the
simulation source code, so I just created an Orientation to keep track
of it in the control room.

Issue #214